E_Project5_HeartDisease
This database contains 14 attributes . In particular, the Cleveland database is the only one that has been used by ML researchers to this date. The "goal" field refers to the presence of heart disease in the patient. It is integer valued from 0 (no presence) to 4. Experiments with the Cleveland database have concentrated on simply attempting to distinguish presence (values 1,2,3,4) from absence (value 0).
Here we are using classification algorithm. For classification we are using 3rd column as the target data.
Its used to distinguish presence (values 1,2,3,4) from absence value 0.